refactor: split apart existing Nav UI into smaller, more manageable pieces
