+++ packages/uikit-workshop/src/scripts/components/pl-nav/src/NavItem.js
@@ -1,0 +1,8 @@
+import { h } from 'preact';
+import cx from 'classnames';
+
+export const NavItem = props => {
+  const classes = cx('pl-c-nav__item', props.className);
+
+  return <li className={classes}>{props.children}</li>;
+};
