+++ packages/uikit-workshop/src/scripts/components/pl-nav/src/PatternState.js
@@ -1,0 +1,10 @@
+import { h } from 'preact';
+
+export const PatternState = props => {
+  return (
+    <span
+      class={`pl-c-pattern-state pl-c-pattern-state--${props.variant}`}
+      title={props.variant}
+    />
+  );
+};
